// frontend/src/app/auth/interfaces/interfaces.ts
export interface AuthResponse{
    idUsuario?         : number;
    correo?            : string;
    ok                 : boolean;
    idPersona?         : number;
    nombres?           : string;
    apellidos?         : string
    identificacion?    : string;
    foto?              : string;
    tipoIdentificacion?: string
    nombreRol?         : string
    idRol?             : number;
    token?             : string;
    error?             : {
      msg              : string
    }
}
export interface Usuario{
  idUsuario?         : number;
  correo?            : string;
  idPersona?         : number;
  nombres?           : string;
  apellidos?         : string
  identificacion?    : string;
  foto?              : string;
  tipoIdentificacion?: string
  nombreRol?         : string
  idRol?             : number;
  rol?               : string;
}
export interface TokenResponse{
  ok:  boolean;
  msg: string;
}
export interface TokenPayload{
  exp?     : number;
  iat?     : number;
  rol?     : string;
  username?: string;
}
export interface Curso {
  id?         : number;
  nombre      : string;
  descripcion : string;
}

export interface Cursos {
  curso   : Curso[];
  total   : number;
  status  : string;
}

export interface Programas {
  status:    string;
  programas: Programa[];
  total:     number;
}

export interface Programa {
  id?:     number;
  nombre?: string;
  status?: string;
}
export interface Respuesta {
  status  :  number;
  message : string;
  error?   : string;
  foto?    : string;
}
export interface Asociacion {
  curso?           : number;
  programa?        : number;
}
export interface Asociaciones {
  asociaciones : Asociacion[];
  total        : number;
  status       : string;
}
export interface CursoAsociacion {
  idCursoPrograma? : number;
  idPrograma?      : number;
  id?              : number;//id curso
}
export interface Personas {
  status:   number;
  personas: Persona[];
  total:    number;
}
export interface EditCursoAsociacion {
  curso?    : number;
  programa? : number;
  id?       : number;//id curso
}
export interface Persona {
  id?:             number;
  tipoId:          string;
  idTipo:          number;
  identificacion:  string;
  nombres?:        string | null;
  apellidos?:      string | null;
  celular?:        string | null;
  status?:         string;
  programa?:       number;
}

export interface Matriculas {
  status:     string;
  matriculas: Matricula[];
  total:      number;
}

export interface Matricula {
  id?:          number;
  status?:      string;
  cursoNombre?: string;
  curso?:       number;
  estudiante?:  string;
  usuario?:     number;
  identificacion?: string;
}

// Generated by https://quicktype.io

export interface Usuarios {
  status:   string;
  usuarios: Usuario[];
  total:    number;
}
// Generated by https://quicktype.io

export interface UsuarioLogin {
  persona:  number;
  rol:      number;
  username: string;
  password: string;
  correo:   string;
}

export interface UsuarioNuevo {
  id?:       number;
  identificacion:  number;
  persona:  number;
  rol?:      number;
  username: string;
  password: string;
  correo:   string;
}

export interface Contenidos {
  status:     string;
  contenidos: Contenido[];
  total:      number;
}

export interface Contenido {
  id:          number;
  tema:        string;
  url:         string;
  status:      string;
  cursoNombre: string;
  curso:       number;
  descripcion: string;
}
export interface eliminarContenido {
  id:                    number;
  idDocumento:           number;
  idContenidoDocumento:  number;
}
export interface actualizarCorreoFoto{
  correo : string;
  id     : number,
  foto   : any
}